+++ flo-server/src/lib.rs
@@ -27,7 +27,7 @@ extern crate tempdir;
 pub mod logging;
 pub mod server;
 pub mod embedded;
-pub mod new_engine;
+pub mod engine;
 pub mod event_loops;
 pub mod channels;
 pub mod atomics;
